Create .gitignore
A .gitignore file is used to stop Git from tracking files you don't want it to track. The usual things put in there are intermediate files (like object files), executables, and hidden OS files (like desktop.ini in Windows, or .DS_store on Mac OS). Since we're making a website, I don't know if there's anything else we'd need to ignore, but this file can be edited as needed.